Dynamically pad content based on size of header
[Fixes #161797234]
