+++ package.json
@@ -64,6 +64,7 @@
     "react-dom": "^16.4.2",
     "react-hot-loader": "^4.3.4",
     "react-redux": "^5.0.7",
+    "react-resize-detector": "^3.2.0",
     "react-router": "^4.3.1",
     "react-static": "^5.9.12",
     "react-time-ago": "^3.0.3",
